+++ zcm/cli/list.py
@@ -41,15 +41,16 @@ class List:
                             help='zfs filesystem or path to show')
 
     def __init__(self, options):
-        manager = Manager(options.path)
         table = []
-        for clone in manager.clones:
-            table.append({
-                'a': '*' if manager.active == clone else ' ',
-                'id': clone['id'],
-                'mountpoint': clone['mountpoint'],
-                'origin': clone['origin_id'] if clone['origin_id'] else '',
-                'date': datetime.fromtimestamp(clone['creation']),
-                'size': format_bytes(clone['used'])
-            })
+        for manager in Manager.get_managers():
+            for clone in manager.clones:
+                table.append({
+                    'name': manager.name,
+                    'a': '*' if manager.active == clone else ' ',
+                    'id': clone['id'],
+                    'mountpoint': clone['mountpoint'],
+                    'origin': clone['origin_id'] if clone['origin_id'] else '',
+                    'date': datetime.fromtimestamp(clone['creation']),
+                    'size': format_bytes(clone['used'])
+                })
         print_table(table, truncate=(not options.no_trunc))